refactor(repeaters): naming and grouping improvements
relates-to #22
