--- packages/repeater/src/handlers/index.ts
@@ -1,2 +1,0 @@
-export * from './events';
-export { ExecuteRequestHandler } from './ExecuteRequestHandler';
