+++ packages/repeater/src/lib/index.ts
@@ -1,5 +1,4 @@
 export { Repeater } from './Repeater';
 export { RepeaterFactory } from './RepeaterFactory';
 export { RepeaterOptions } from './RepeaterOptions';
-export { RepeaterStatus } from './RepeaterStatus';
-export { RequestExecutorOptions } from '../request-executor/RequestExecutorOptions';
+export { RepeaterStatus } from '../models/RepeaterStatus';
