+++ packages/repeater/src/request-runner/protocols/index.ts
@@ -1,0 +1,2 @@
+export * from './HttpRequestRunner';
+export * from './WsRequestRunner';
